Eliminated borders and added buttons
I was spending WAY to much time trying to make the etcher work with a
border enabled. The geometry of the problem changes when working with
widths which are not evenly divisable by the desired number of columns.
The end result was a gap between the etch cells and the border, so I
finally gave up. This approach just plops the etcher onto the frame and
that is that.